+++ src/configs/markdown-preferences.ts
@@ -82,12 +82,13 @@ export const markdownPreferencesUnConfig: UnConfigFn<'markdownPreferences'> = as
 
   const defaultPreserveWords = getKeysOfTruthyValues({
     ...Object.fromEntries(
-      (markdownPreferencesPlugin?.resources.defaultPreserveWords || []).map((defaultWord) => [
-        defaultWord,
-        true,
-      ]),
+      [
+        ...markdownPreferencesPlugin.resources.defaultPreserveWords,
+        // Remove once they're added: https://github.com/ota-meshi/eslint-plugin-markdown-preferences/issues/173
+        'I',
+        'MIT',
+      ].map((defaultWord) => [defaultWord, true]),
     ),
-    ['I' as string]: true, // Fixes an issue that 'I' is suggested to be lowercased when sentence case is preferred
     ...(!Array.isArray(wordsToPreserveCasingOf) && wordsToPreserveCasingOf),
   });
   if (Array.isArray(wordsToPreserveCasingOf)) {